just some small refactoring
